+++ frontend/src/data/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added "Force 480p previews on HEVC" for users who are having issues with HEVC videos not previewing
 - Added option "Unselect all after exports", on by default
 - Fixed clips grid resetting to the top on export
+- Fixed grid columns being seperated and unresponsive
 
 ## v1.2.6 - v1.2.8
 - Fixed update(s) not applying correctly (bumping version)